revise update method to allow breaking out; implement object diff
Update now works as a two stage process. By default it updates and
commits changes to the repo but it can be paused by setting commit =
false. The pause happens just before the deposit object is moved back
to the repo so the internal state is set to the object in the
deposit path and all methods operate relative to that version. The user
is responsible committing or cleaning up.

This update also includes a method to diff version content.

commit 972b186e8346986dc6b4b97d99ccb0620fd2d8b0
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Sep 11 15:01:48 2020 +1000

    add tests to check object diff

commit 0fadcf7465dc29ad7b0f2a750a110d5b1de2bc5f
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Sep 11 15:01:29 2020 +1000

    implement object diff and cleanup

commit 296c1d1ad9170759b1c98ddd167916208f85030c
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Sep 11 13:44:38 2020 +1000

    refactor update / commit to break out after versioning

commit a083fe1f1f6879045c49006b3de8dc4c83e51381
Author: Marco La Rosa <m@lr.id.au>
Date:   Fri Sep 11 12:07:34 2020 +1000

    put a break into update

    Split update into an update and a commit phase such that the caller can
    choose to break out out of the update loop before the commit to the
    repository occurs.